Update Debian.dockerfile base image
Updated from bullseye to bookworm. 
Not a critical upgrade as bullseye still gets security updates, but this is for the future. In two years, when I probably do my next re-install, bookworm is the more up-to-date choise, so when I forget to check the base image, this has it mostly taken care of.